Refactor[submodule] #2: Initialize controller & validator of Auth
